+++ sixpack/config.py
@@ -18,8 +18,9 @@ else:
         'redis_host': os.environ.get('SIXPACK_CONFIG_REDIS_HOST', "localhost"),
         'redis_password': os.environ.get('SIXPACK_CONFIG_REDIS_PASSWORD', None),
         'redis_prefix': os.environ.get('SIXPACK_CONFIG_REDIS_PREFIX', "sxp"),
+        'redis_socket_timeout': os.environ.get('SIXPACK_CONFIG_REDIS_SOCKET_TIMEOUT', None),
+        'redis_sentinel_service_name': os.environ.get('SIXPACK_CONFIG_REDIS_SENTINEL_SERVICE_NAME', None),
         'redis_db': int(os.environ.get('SIXPACK_CONFIG_REDIS_DB', '15')),
-        'enable_whiplash': to_bool(os.environ.get('SIXPACK_CONFIG_WHIPLASH', 'False')),
         'robot_regex': os.environ.get('SIXPACK_CONFIG_ROBOT_REGEX', "$^|trivial|facebook|MetaURI|butterfly|google|"
                                                                     "amazon|goldfire|sleuth|xenu|msnbot|SiteUptime|"
                                                                     "Slurp|WordPress|ZIBB|ZyBorg|pingdom|bot|yahoo|"
@@ -33,3 +34,10 @@ else:
             'exchange': os.environ.get('SIXPACK_CONFIG_RABBIT_EXCHANGE', None)
         }
     }
+
+    if 'SIXPACK_CONFIG_REDIS_SENTINELS' in os.environ:
+        sentinels = []
+        for sentinel in os.environ['SIXPACK_CONFIG_REDIS_SENTINELS'].split(","):
+            server,port = sentinel.split(":")
+            sentinels.append([server, int(port)])
+        CONFIG['redis_sentinels'] = sentinels