+++ sixpack/server.py
@@ -10,7 +10,10 @@ from werkzeug.routing import Map, Rule
 from werkzeug.exceptions import HTTPException, NotFound
 
 from . import __version__
+from api import participate, convert
+
 from config import CONFIG as cfg
+from utils import to_bool
 
 try:
     import db
@@ -28,7 +31,7 @@ from utils import service_unavailable_on_connection_error, json_error, json_succ
 
 class Sixpack(object):
 
-    def __init__(self, redis_conn, queue = None):
+    def __init__(self, redis_conn, queue=None):
         self.redis = redis_conn
 
         self.config = cfg
@@ -38,6 +41,7 @@ class Sixpack(object):
             Rule('/_status', endpoint='status'),
             Rule('/participate', endpoint='participate'),
             Rule('/convert', endpoint='convert'),
+            Rule('/experiments/<name>', endpoint='experiment_details'),
             Rule('/favicon.ico', endpoint='favicon')
         ])
 
@@ -97,26 +101,21 @@ class Sixpack(object):
         if client_id is None or experiment_name is None:
             return json_error({'message': 'missing arguments'}, request, 400)
 
-        client = Client(client_id, self.redis)
+        dt = None
+        if request.args.get("datetime"):
+            dt = dateutil.parser.parse(request.args.get("datetime"))
 
         try:
-            experiment = Experiment.find(experiment_name, self.redis, self.queue)
-            if cfg.get('enabled', True):
-                dt = None
-                if request.args.get("datetime"):
-                    dt = dateutil.parser.parse(request.args.get("datetime"))
-                alternative = experiment.convert(client, dt=dt, kpi=kpi)
-            else:
-                alternative = experiment.control.name
+            alt = convert(experiment_name, client_id, kpi=kpi, datetime=dt, redis=self.redis, queue=self.queue)
         except ValueError as e:
             return json_error({'message': str(e)}, request, 400)
 
         resp = {
             'alternative': {
-                'name': alternative
+                'name': alt.name
             },
             'experiment': {
-                'name': experiment.name,
+                'name': alt.experiment.name,
             },
             'conversion': {
                 'value': None,
@@ -129,51 +128,54 @@ class Sixpack(object):
 
     @service_unavailable_on_connection_error
     def on_participate(self, request):
-        opts = {}
         alts = request.args.getlist('alternatives')
         experiment_name = request.args.get('experiment')
         force = request.args.get('force')
         client_id = request.args.get('client_id')
-        traffic_fraction = request.args.get('traffic_fraction', 1)
+        traffic_fraction = float(request.args.get('traffic_fraction', 1))
+        prefetch = to_bool(request.args.get('prefetch', 'false'))
 
         if client_id is None or experiment_name is None or alts is None:
             return json_error({'message': 'missing arguments'}, request, 400)
 
-        opts['traffic_fraction'] = traffic_fraction
+        dt = None
+        if request.args.get("datetime"):
+            dt = dateutil.parser.parse(request.args.get("datetime"))
 
-        try:
-            experiment = Experiment.find_or_create(experiment_name, alts, self.redis, opts, self.queue)
-        except ValueError as e:
-            return json_error({'message': str(e)}, request, 400)
-
-        alternative = None
-        if force and force in alts:
-            alternative = force
-        elif not cfg.get('enabled', True):
-            alternative = alts[0]
-        elif experiment.winner is not None:
-            alternative = experiment.winner
-        elif should_exclude_visitor(request):
-            alternative = alts[0]
+        if should_exclude_visitor(request):
+            exp = Experiment.find(experiment_name, redis=self.redis, queue=sele.queue)
+            if exp.winner is not None:
+                alt = exp.winner
+            else:
+                alt = exp.control
         else:
-            dt = None
-            if request.args.get("datetime"):
-                dt = dateutil.parser.parse(request.args.get("datetime"))
-            client = Client(client_id, self.redis)
-            alternative = experiment.get_alternative(client, dt=dt).name
+            try:
+                alt = participate(experiment_name, alts, client_id,
+                                  force=force, traffic_fraction=traffic_fraction,
+                                  prefetch=prefetch, datetime=dt, redis=self.redis, queue=self.queue)
+            except ValueError as e:
+                return json_error({'message': str(e)}, request, 400)
 
         resp = {
             'alternative': {
-                'name': alternative
+                'name': alt.name
             },
             'experiment': {
-                'name': experiment.name,
+                'name': alt.experiment.name,
             },
             'client_id': client_id,
             'status': 'ok'
         }
 
         return json_success(resp, request)
+
+    @service_unavailable_on_connection_error
+    def on_experiment_details(self, request, name):
+        exp = Experiment.find(name, redis=self.redis)
+        if exp is None:
+            return json_error({'message': 'experiment not found'}, request, 404)
+
+        return json_success(exp.objectify_by_period('day', True), request)
 
 
 def should_exclude_visitor(request):