+++ sixpack/static/js/experiment.js
@@ -7,18 +7,27 @@ $(function () {
     _.templateSettings.variable = 'experiment';
 
     my.el = el;
-    my.name = name;
+    my.codedName = name;
+    my.name = name.match(/\w+/g).join('-');
+    my.encoded = encodeURIComponent(name);
     my.callback = callback;
 
     my.template = _.template($('#experiment-template').html());
 
     my.getData = function (callback) {
-      var url = '/experiments/' + my.name + '.json?period=day';
+      var url = '/experiments/' + my.encoded + '.json?period=day';
       if (typeof kpi != 'undefined' && kpi !== false) {
         url += '&kpi=' + kpi;
       }
-      $.getJSON(url, function (data) {
+      
+      var promise = $.getJSON(url);
+
+      promise.done(function(data) {
         callback(data);
+      });
+
+      promise.fail(function(resp) {
+        $(my.el).trigger('fail', [resp]);
       });
     };
 
@@ -45,7 +54,7 @@ $(function () {
 
       my.el.append(my.template(data));
 
-      $("li[data-name='" + my.name + "'] tr").on({
+      $("li[data-name='" + my.codedName + "'] tr").on({
         mouseover: function () {
           var alt_name = $(this).attr('class');
           if (!alt_name) return;