+++ sixpack/web.py
@@ -1,3 +1,5 @@
+import urllib
+
 from flask import Flask
 from flask import render_template, abort, request, url_for, redirect, jsonify, make_response
 from flask.ext.seasurf import SeaSurf
@@ -9,7 +11,10 @@ from . import __version__
 from config import CONFIG as cfg
 import db
 from models import Experiment
+from analysis import ExportExperiment
 import utils
+
+import re
 
 app = Flask(__name__)
 csrf = SeaSurf(app)
@@ -18,12 +23,10 @@ js = Bundle('js/vendor/jquery.js', 'js/vendor/d3.js',
             'js/vendor/bootstrap.js', 'js/experiment.js', 'js/chart.js',
             'js/sixpack.js', 'js/vendor/underscore-min.js', 'js/vendor/spin.min.js',
             'js/vendor/waypoints.min.js', 'js/vendor/zeroclipboard.min.js',
-            filters=['closure_js'],
             output="{0}/sixpack.js".format(cfg.get('asset_path', 'gen')))
 
 css = Bundle('css/vendor/bootstrap.css',
              'css/vendor/bootstrap-responsive.css', 'css/sixpack.css',
-             filters=['yui_css'],
              output="{0}/sixpack.css".format(cfg.get('asset_path', 'gen')))
 
 assets = Environment(app)
@@ -40,21 +43,21 @@ def status():
 
 @app.route("/")
 def hello():
-    experiments = Experiment.all(db.REDIS)
+    experiments = Experiment.all(exclude_archived=True, redis=db.REDIS)
     experiments = [exp.name for exp in experiments]
     return render_template('dashboard.html', experiments=experiments, page='home')
 
 
 @app.route('/archived')
 def archived():
-    experiments = Experiment.all(db.REDIS, False)
+    experiments = Experiment.all(exclude_archived=False, redis=db.REDIS)
     experiments = [exp.name for exp in experiments if exp.is_archived()]
     return render_template('dashboard.html', experiments=experiments, page='archived')
 
 
 @app.route('/experiments.json')
 def experiment_list():
-    experiments = Experiment.all(db.REDIS)
+    experiments = Experiment.all(redis=db.REDIS)
     period = determine_period()
     experiments = [simple_markdown(exp.objectify_by_period(period)) for exp in experiments]
     return jsonify({'experiments': experiments})
@@ -79,7 +82,8 @@ def json_details(experiment_name):
 def export(experiment_name):
     experiment = find_or_404(experiment_name)
 
-    response = make_response(experiment.csvify())
+    export = ExportExperiment(experiment=experiment)
+    response = make_response(export())
     response.headers["Content-Type"] = "text/csv"
     # force a download with the content-disposition headers
     filename = "sixpack_export_{0}".format(experiment_name)
@@ -162,6 +166,7 @@ def internal_server_error(e):
 
 def find_or_404(experiment_name):
     try:
+        experiment_name = url=urllib.unquote(experiment_name).decode('utf8') 
         exp = Experiment.find(experiment_name, db.REDIS)
         if request.args.get('kpi'):
             exp.set_kpi(request.args.get('kpi'))
@@ -184,10 +189,11 @@ def simple_markdown(experiment):
         experiment['pretty_description'] = markdown(description)
     return experiment
 
-
 app.secret_key = cfg.get('secret_key')
 app.jinja_env.filters['number_to_percent'] = utils.number_to_percent
 app.jinja_env.filters['number_format'] = utils.number_format
+app.jinja_env.filters['sanitize'] = utils.sanitize_experiment
+app.jinja_env.filters['regex_replace'] = utils.regex_replace
 toolbar = DebugToolbarExtension(app)